refactors tests for base classes
