+++ tests/tests_approximators/test_approximator_kernelshap.py
@@ -1,7 +1,5 @@
 """This test module contains all tests regarding the SV KernelSHAP regression approximator."""
 
-from copy import copy, deepcopy
-
 import numpy as np
 import pytest
 
@@ -19,18 +17,6 @@ def test_initialization(n):
     assert approximator.top_order is False
     assert approximator.min_order == 0
     assert approximator.iteration_cost == 1
-
-    approximator_copy = copy(approximator)
-    approximator_deepcopy = deepcopy(approximator)
-    approximator_deepcopy.index = "something"
-    assert approximator_copy == approximator  # check that the copy is equal
-    assert approximator_deepcopy != approximator  # check that the deepcopy is not equal
-    approximator_string = str(approximator)
-    assert repr(approximator) == approximator_string
-    assert hash(approximator) == hash(approximator_copy)
-    assert hash(approximator) != hash(approximator_deepcopy)
-    with pytest.raises(ValueError):
-        _ = approximator == 1
 
 
 @pytest.mark.parametrize("n, budget", [(7, 380), (7, 380), (7, 100)])
