+++ tests/tests_approximators/test_approximator_permutation_sii.py
@@ -1,7 +1,5 @@
 """This test module contains all tests regarding the SII permutation sampling approximator."""
 
-from copy import copy, deepcopy
-
 import pytest
 
 from shapiq.approximator.permutation import PermutationSamplingSII
@@ -34,18 +32,6 @@ def test_initialization(n, max_order, top_order, index, expected):
     assert approximator.min_order == (max_order if top_order else 0)
     assert approximator.iteration_cost == expected
     assert approximator.index == index
-
-    approximator_copy = copy(approximator)
-    approximator_deepcopy = deepcopy(approximator)
-    approximator_deepcopy.index = "something"
-    assert approximator_copy == approximator  # check that the copy is equal
-    assert approximator_deepcopy != approximator  # check that the deepcopy is not equal
-    approximator_string = str(approximator)
-    assert repr(approximator) == approximator_string
-    assert hash(approximator) == hash(approximator_copy)
-    assert hash(approximator) != hash(approximator_deepcopy)
-    with pytest.raises(ValueError):
-        _ = approximator == 1
 
 
 @pytest.mark.parametrize("index", ["SII", "k-SII"])
