+++ tests/tests_explainer/test_explainer_base_functionality.py
@@ -1,0 +1,12 @@
+"""This module tests the base functionality of the explainer class."""
+
+import pytest
+
+from shapiq import Explainer
+
+
+def test_explainer():
+    """Tests if the attributes and properties of explainers are set correctly."""
+
+    with pytest.raises(TypeError):
+        Explainer()
